chore: extract helpers to separate files
